refactor: improve type safety by using static_cast for argument conversions in JitExecutor and AsmComplexOperationManager, and update PushPopOptimiser to use size_t for index pairs

+++ jit/oil-to-asm-realisation/optimisers/PushPopOptimiser.cpp
@@ -6,7 +6,7 @@ std::vector<AssemblyInstruction> optimize_push_pop_pairs(
   const std::vector<AssemblyInstruction>& instructions) {
   
   std::vector<AssemblyInstruction> optimized;
-  std::vector<std::pair<int, int>> push_pop_pairs; // пары индексов (push, pop)
+  std::vector<std::pair<size_t, size_t>> push_pop_pairs; // пары индексов (push, pop)
   
   // Сначала находим все пары PUSH RAX / POP RAX
   for (size_t i = 0; i < instructions.size(); ++i) {
